Extract migrateFileMeta into separate file called cloneFileMeta
